+++ src/main/java/kitchenpos/dto/request/OrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 import java.util.List;
 import java.util.stream.Collectors;
