+++ src/main/java/kitchenpos/dto/response/MenuGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.response;
 
 public class MenuGroupResponse {
