+++ src/main/java/kitchenpos/dto/response/MenuResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.response;
 
 import java.math.BigDecimal;
 import java.util.List;
